Fix: Minebot command text for activating lights (#90799)
## About The Pull Request

Due to the wrong path being specified, if you try to order the minebot
to turn on its lights, it'll talk about dumping ore instead. The actual
functionality is fine, but the feedback isn't. This fixes that.

## Why It's Good For The Game

The current bug just makes things confusing. Fewer sources of confusion
in this game is better.

## Changelog

:cl:
fix: Made it so minebot commands don't talk about dumping ore when
you're toggling the lights.
/:cl: