Being intoxicated has a chance for you to not register the trauma of surgery based on intoxication. Puts some whiskey in Tram and Meta's robotics lab. For that authentic back alley surgeon feel. (#90956)
## About The Pull Request

Based on how intoxicated you are, it is possible for you to avoid
obtaining moodlets and screaming from being cut open during surgery.

Adds some whiskey to the Tram and Meta robotics labs, as well as a shot
glass to drink it.

## Why It's Good For The Game

Alcohol is a valid anesthetic in this game, but the only way through
which you can avoid these moodlets is actually going fully under, so
alcohol is less ideal by comparison.

Having the option of liquoring up before surgery to try and tough out
the trauma is very on theme with the games themes of medical
malpractice.

On that note, roboticists using alcohol rather than proper anesthetic is
equally on brand. I just want to see a robo pessimistically pouring
their patient and themself a drink before picking up a scalpel.

## Changelog
:cl:
qol: Being intoxicated enough can potentially beat back the trauma of
surgery. It's not reliable.
qol: Robotics on Tram and Metastation have a bottle of whiskey. You
know, for taking the edge off.
/:cl: