Stops you getting into a mech if you have stuff in your hands at all. (#91046)
## About The Pull Request

Makes it so that it is impossible to enter a mech if your hands are
full.

## Why It's Good For The Game

This can result in some...very unintended behavior should somehow you
cannot be forced to drop items in your hands.

## Changelog
:cl:
fix: You must have both your hands empty before entering a mech. (But
items which replace your hand, like the light later or the hand of god,
won't prevent you from getting in)
/:cl: