changes minimum age from 17 to 18 (#91200)
## About The Pull Request

does what the title says

## Why It's Good For The Game

people just use this to be creepy. we require people to be at least 18
to play, why would we intentionally let someone roleplay as underage.
why would nanotrasen hire a 17 year old??? recent events on the station
have shown that keeping this around is a bad idea.

## Changelog

:cl:
config: The minimum age on all Nanotrasen research stations is now 18.
/:cl: