Narcolepsy quirk (#91184)
## About The Pull Request

This is a PR I made a while ago for downstreams, and it has become my
favorite negative quirk to play with. It grants your character a
slightly more forgiving version of the narcolepsy trauma, as it has
lower odds of making you snooze plus a bonus list of reagents that will
suppress symptoms (`modafinil` and `synaptizine` for now), and cares
about if you are caffeinated, (have `trait_simulated`).

The rate at which you fall asleep/get drowsy is appropriate in my
opinion, though I could understand if players would want it to be more
common.


![image](https://github.com/user-attachments/assets/4a2a6f3e-0a4b-4085-9d43-acf28778d991)

## Why It's Good For The Game

Really fun quirk to play with, adds more risk to your gameplay. Use your
over-the-counter stimulants to try and win from a stun baton.

## Changelog


:cl:
add: Adds Narcolepsy as a -8 negative quirk!
refactor: Slightly adjusts the logic of the narcolepsy trauma
balance: The narcolepsy trauma its symptoms can now be softened by being
caffeinated, and modafinil or synaptizine suppress it entirely
/:cl: