Expands the medical foyer on Nebulastation by simply removing some doors, forming a large donut-shaped reception area. (#91391)
## About The Pull Request

Basically what it says on the tin, but also includes removing the access
restrictions to a few additional side seating area doors.

## Why It's Good For The Game

The most frustrating part of Nebulastation's medical area is the
difficulty to get access into the deeper parts of the department to get
seen by medical doctors. Medical doctors often need to go quite a
distance to check who is or isn't at the front desk, or to see if
someone has been dropped off for medical attention. This is frustrating
on both sides.

There isn't much reason for medical doctors to be out in that wider area
most of the time. It also doesn't contain much of value in the first
place. Anything critical is already in the main front desk.

A lot of the seating area is mostly useless and just airgaps segments of
medical even further. Making them more broadly accessible helps with
visibility further if people choose to use those areas. Or, in the case
of the psychologist, actually get seen at all because their main office
is buried in medical in a near unremarkable corner. (Not that this is a
Nebula specific problem)

## Changelog
:cl:
map: Opens up sections of Nebulastation's medical department to allow
for greater visibility into and out of medical.
/:cl: