Fix two runtimes related to decal lazy init reagents (#91503)
## About The Pull Request

#91054 changed decals so they no longer init reagent holders on
`Initialize`, which means these `expose_turf` reactions were runtiming
because the reagent holder didn't exist when they tried to
`add_reagent`. This rewrites their code to make sure init is handled,
but also to take advantage of the fact that there won't be existing
reagents (like in the case of ants where it does - 5). It was also
necessary to add code to increase the max volume of the reagent holder
because `lazy_init_reagents` sets max volume to `reagent_amount` instead
of 300 like the old code

## Why It's Good For The Game

Bug fix

## Changelog
:cl:
fix: fixed runtimes with splashing ants and uranium onto tiles
/:cl: